requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.11
django-allauth==0.61.1
django-crispy-forms==2.1
filelock==3.13.1
idna==3.6
oauthlib==3.2.2
platformdirs==4.2.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.4.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.1
